feat(ODB-Lite): Support list type value.Add command "addr(key,value)" and "addl(key,value)"
To create a list.you can use command like this： set(key,[value1;value2])
Remember：
List use a full pair of "[]" to be identified.
Values in list using a ';' to separate.

Use "addr(key,value)" to add a value to list'right
Use "addl(key,value)" to add a value to list'left